Added a middle mouse button click for debug to view current state of the board to ensure backend data is matching front end.
Removed color logic related to the UI from the grid as the intenion is that the grid is portable and usable in different projects. Instead, grid now has a single dictionary organized by colors that have values of either a cell (tuple) or a set of cells. Based on the information in the dictionary, grid can color cells without needing to know anything about the logic of cell coloring decisions.